Effects.Concurrent.Thread

Description

Provides the MonadThread typeclass.

Since: 0.1

Thread Effect

class Monad m => MonadThread (m :: Type -> Type) where

Represents thread effects.

Since: 0.1

Methods

threadDelay :: Int -> m ()

Lifted threadDelay.

Since: 0.1

throwTo :: (Exception e, HasCallStack) => ThreadId -> e -> m ()

Lifted throwTo.

Since: 0.1

getNumCapabilities :: m Int

Lifted getNumCapabilities.

Since: 0.1

setNumCapabilities :: Int -> m ()

Lifted setNumCapabilities.

Since: 0.1

threadCapability :: ThreadId -> m (Int, Bool)

Lifted threadCapability.

Since: 0.1

myThreadId :: m ThreadId

Lifted myThreadId.

Since: 0.1

labelThread :: ThreadId -> String -> m ()

Lifted labelThread.

Since: 0.1

threadLabel :: ThreadId -> m (Maybe String)

Lifted threadLabel.

Since: 0.1

microsleep :: (HasCallStack, MonadThread m) => Natural -> m ()

threadDelay in terms of unbounded Natural rather than Int i.e. runs sleep in the current thread for the specified number of microseconds.

Since: 0.1

sleep :: (HasCallStack, MonadThread m) => Natural -> m ()

Runs sleep in the current thread for the specified number of seconds.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

data ThreadId

A ThreadId is an abstract type representing a handle to a thread. ThreadId is an instance of Eq, Ord and Show, where the Ord instance implements an arbitrary total ordering over ThreadIds. The Show instance lets you convert an arbitrary-valued ThreadId to string form; showing a ThreadId value is occasionally useful when debugging or diagnosing the behaviour of a concurrent program.

Note: in GHC, if you have a ThreadId, you essentially have a pointer to the thread itself. This means the thread itself can't be garbage collected until you drop the ThreadId. This misfeature would be difficult to correct while continuing to support threadStatus.

QSem Effect

class Monad m => MonadQSem (m :: Type -> Type) where

Effect for QSem semaphore.

Since: 0.1

Methods

newQSem :: Int -> m QSem

Lifted newQSem.

Since: 0.1

waitQSem :: QSem -> m ()

Lifted waitQSem.

Since: 0.1

signalQSem :: QSem -> m ()

Lifted signalQSem.

Since: 0.1

class Monad m => MonadQSemN (m :: Type -> Type) where

Effect for QSemN semaphore.

Since: 0.1

Methods

newQSemN :: Int -> m QSemN

Lifted newQSemN.

Since: 0.1

waitQSemN :: QSemN -> Int -> m ()

Lifted waitQSemN.

Since: 0.1

signalQSemN :: QSemN -> Int -> m ()

Lifted signalQSemN.

Since: 0.1

Reexports

data QSem

QSem is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSem calls.

The pattern

  bracket_ waitQSem signalQSem (...)

is safe; it never loses a unit of the resource.

data QSemN

QSemN is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSemN calls.

The pattern

  bracket_ (waitQSemN n) (signalQSemN n) (...)

is safe; it never loses any of the resource.