Effects.Concurrent.Thread

Description

Provides the MonadThread typeclass.

Since: 0.1

Thread Effect

class Monad m => MonadThread m where

Represents thread effects.

Since: 0.1

Methods

threadDelay :: HasCallStack => Int -> m ()

Suspends the current thread for a given number of microseconds (GHC only).

There is no guarantee that the thread will be rescheduled promptly when the delay has expired, but the thread will never continue to run earlier than specified.

Since: 0.1

throwTo :: (Exception e, HasCallStack) => ThreadId -> e -> m ()

throwTo raises an arbitrary exception in the target thread (GHC only).

Exception delivery synchronizes between the source and target thread: throwTo does not return until the exception has been raised in the target thread. The calling thread can thus be certain that the target thread has received the exception. Exception delivery is also atomic with respect to other exceptions. Atomicity is a useful property to have when dealing with race conditions: e.g. if there are two threads that can kill each other, it is guaranteed that only one of the threads will get to kill the other.

Whatever work the target thread was doing when the exception was raised is not lost: the computation is suspended until required by another thread.

If the target thread is currently making a foreign call, then the exception will not be raised (and hence throwTo will not return) until the call has completed. This is the case regardless of whether the call is inside a mask or not. However, in GHC a foreign call can be annotated as interruptible, in which case a throwTo will cause the RTS to attempt to cause the call to return; see the GHC documentation for more details.

Important note: the behaviour of throwTo differs from that described in the paper "Asynchronous exceptions in Haskell" (http://research.microsoft.com/~simonpj/Papers/asynch-exns.htm). In the paper, throwTo is non-blocking; but the library implementation adopts a more synchronous design in which throwTo does not return until the exception is received by the target thread. The trade-off is discussed in Section 9 of the paper. Like any blocking operation, throwTo is therefore interruptible (see Section 5.3 of the paper). Unlike other interruptible operations, however, throwTo is always interruptible, even if it does not actually block.

There is no guarantee that the exception will be delivered promptly, although the runtime will endeavour to ensure that arbitrary delays don't occur. In GHC, an exception can only be raised when a thread reaches a safe point, where a safe point is where memory allocation occurs. Some loops do not perform any memory allocation inside the loop and therefore cannot be interrupted by a throwTo.

If the target of throwTo is the calling thread, then the behaviour is the same as throwIO, except that the exception is thrown as an asynchronous exception. This means that if there is an enclosing pure computation, which would be the case if the current IO operation is inside unsafePerformIO or unsafeInterleaveIO, that computation is not permanently replaced by the exception, but is suspended as if it had received an asynchronous exception.

Note that if throwTo is called with the current thread as the target, the exception will be thrown even if the thread is currently inside mask or uninterruptibleMask.

Since: 0.1

getNumCapabilities :: HasCallStack => m Int

Returns the number of Haskell threads that can run truly simultaneously (on separate physical processors) at any given time. To change this value, use setNumCapabilities.

Since: 0.1

setNumCapabilities :: HasCallStack => Int -> m ()

Set the number of Haskell threads that can run truly simultaneously (on separate physical processors) at any given time. The number passed to forkOn is interpreted modulo this value. The initial value is given by the +RTS -N runtime flag.

This is also the number of threads that will participate in parallel garbage collection. It is strongly recommended that the number of capabilities is not set larger than the number of physical processor cores, and it may often be beneficial to leave one or more cores free to avoid contention with other processes in the machine.

Since: 0.1

threadCapability :: HasCallStack => ThreadId -> m (Int, Bool)

Returns the number of the capability on which the thread is currently running, and a boolean indicating whether the thread is locked to that capability or not. A thread is locked to a capability if it was created with forkOn.

Since: 0.1

microsleep :: (HasCallStack, MonadThread m) => Natural -> m ()

threadDelay in terms of unbounded Natural rather than Int i.e. runs sleep in the current thread for the specified number of microseconds.

Since: 0.1

sleep :: (HasCallStack, MonadThread m) => Natural -> m ()

Runs sleep in the current thread for the specified number of seconds.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

data ThreadId

A ThreadId is an abstract type representing a handle to a thread. ThreadId is an instance of Eq, Ord and Show, where the Ord instance implements an arbitrary total ordering over ThreadIds. The Show instance lets you convert an arbitrary-valued ThreadId to string form; showing a ThreadId value is occasionally useful when debugging or diagnosing the behaviour of a concurrent program.

Note: in GHC, if you have a ThreadId, you essentially have a pointer to the thread itself. This means the thread itself can't be garbage collected until you drop the ThreadId. This misfeature will hopefully be corrected at a later date.

QSem Effect

class Monad m => MonadQSem m where

Effect for QSem semaphore.

Since: 0.1

Methods

newQSem :: Int -> m QSem

Build a new QSem with a supplied initial quantity. The initial quantity must be at least 0.

Since: 0.1

waitQSem :: QSem -> m ()

Wait for a unit to become available.

Since: 0.1

signalQSem :: QSem -> m ()

Signal that a unit of the QSem is available.

Since: 0.1

class Monad m => MonadQSemN m where

Effect for QSemN semaphore.

Since: 0.1

Methods

newQSemN :: Int -> m QSemN

Build a new QSemN with a supplied initial quantity. The initial quantity must be at least 0.

Since: 0.1

waitQSemN :: QSemN -> Int -> m ()

Wait for the specified quantity to become available.

Since: 0.1

signalQSemN :: QSemN -> Int -> m ()

Signal that a given quantity is now available from the QSemN.

Since: 0.1

Reexports

data QSem

QSem is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSem calls.

The pattern

  bracket_ waitQSem signalQSem (...)

is safe; it never loses a unit of the resource.

data QSemN

QSemN is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSemN calls.

The pattern

  bracket_ (waitQSemN n) (signalQSemN n) (...)

is safe; it never loses any of the resource.